requirement.txt
torch==1.9.0
torchvision==0.10.0
streamlit==1.11.0
Pillow==8.2.0